refactor: package명, package구조 변경
- component package -> componenets package 이름 변경
- controller package 생성
- BaseballGameController class : componenets package -> controller package 이동
