+++ src/test/java/baseball/components/BaseballGameControllerTest.java
@@ -1,5 +1,6 @@
-package baseball.component;
+package baseball.components;
 
+import baseball.controller.BaseballGameController;
 import camp.nextstep.edu.missionutils.test.NsTest;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
