+++ src/test/java/baseball/components/BaseballGameRuleTest.java
@@ -1,4 +1,4 @@
-package baseball.component;
+package baseball.components;
 
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
